refactor(sdk): unify type system and expose customer/subscription resources
