--- src/schemas.ts
@@ -1,1 +1,0 @@
-export * from "@payark/sdk-effect/schemas";
